refactor: split components into smaller files and keep the selected options
