+++ frontend/web/components/base/select/multi-select/index.ts
@@ -1,0 +1,1 @@
+export { MultiSelect } from './MultiSelect'
