+++ frontend/web/components/organisation-settings/usage/OrganisationUsageMetrics.container.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo, useState } from 'react'
 import { Res } from 'common/types/responses'
 import SingleSDKLabelsChart from './components/SingleSDKLabelsChart'
-import MultiSelect from 'components/base/select/MultiSelect'
+import { MultiSelect } from 'components/base/select/multi-select'
 
 export interface OrganisationUsageMetricsProps {
   data?: Res['organisationUsage']
