+++ frontend/web/components/segments/Rule/Rule.tsx
@@ -16,7 +16,6 @@ import {
   splitIfValue,
   isInvalidPercentageSplit,
 } from './utils'
-import { useConditionCache } from './hooks/useConditionCache'
 
 interface RuleProps {
   index: number
@@ -42,8 +41,6 @@ const Rule: React.FC<RuleProps> = ({
 }) => {
   const { conditions: rules } = rule
 
-  const { cacheValue, getCachedValue } = useConditionCache()
-
   const updateCondition = (
     conditionIndex: number,
     updates: Partial<SegmentCondition>,
@@ -59,15 +56,7 @@ const Rule: React.FC<RuleProps> = ({
   }
 
   const setConditionProperty = (conditionIndex: number, property: string) => {
-    const condition = rule.conditions[conditionIndex]
-
-    if (condition.property) {
-      cacheValue(conditionIndex, condition.operator, condition.property, condition.value)
-    }
-
-    const cachedValue = getCachedValue(conditionIndex, condition.operator, property)
-
-    updateCondition(conditionIndex, { property, value: cachedValue })
+    updateCondition(conditionIndex, { property })
   }
 
   const setConditionOperator = (
@@ -82,13 +71,16 @@ const Rule: React.FC<RuleProps> = ({
     )
     const newOperator = operators.find((op) => op.value === operatorValue)
 
-    if (condition.operator && condition.property) {
-      cacheValue(conditionIndex, condition.operator, condition.property, condition.value)
-    }
-
     // Split the append part of the operator as not handled by backend
     const updates: Partial<SegmentCondition> = {
       operator: operatorValue?.split(':')[0],
+    }
+
+    // Clear value when changing to or from IN operator
+    const isChangingToIn = operatorValue === 'IN'
+    const isChangingFromIn = condition?.operator === 'IN'
+    if (isChangingToIn || isChangingFromIn) {
+      updates.value = ''
     }
 
     if (newOperator?.hideValue) {
@@ -116,22 +108,6 @@ const Rule: React.FC<RuleProps> = ({
       } else {
         updates.value = updates.value?.toString().split(':')[0]
       }
-    }
-
-    const finalProperty = updates.property !== undefined ? updates.property : condition.property
-    const finalOperator = operatorValue?.split(':')[0]
-
-    if (finalProperty && !newOperator?.hideValue && operatorValue !== 'PERCENTAGE_SPLIT') {
-      const cachedValue = getCachedValue(conditionIndex, finalOperator, finalProperty)
-
-      if (cachedValue !== '') {
-        // Use cached value if one exists (restore previous input)
-        updates.value = cachedValue
-      } else if (!updates.value) {
-        // No cache and no value set by other logic - clear the input
-        updates.value = ''
-      }
-      // Otherwise: preserve value set by append logic or existing value
     }
 
     updateCondition(conditionIndex, updates)
