+++ frontend/web/components/segments/Rule/components/RuleConditionValueInput.tsx
@@ -7,7 +7,7 @@ import { getDarkMode } from 'project/darkMode'
 import EnvironmentSelectDropdown from './EnvironmentSelectDropdown'
 import TextAreaModal from './TextAreaModal'
 import { checkWhitespaceIssues } from 'components/segments/Rule/utils'
-import MultiSelect from 'components/base/select/MultiSelect'
+import { MultiSelect } from 'components/base/select/multi-select'
 import { safeParseArray } from 'components/segments/Rule/utils/arrayUtils'
 import { useGetEnvironmentsQuery } from 'common/services/useEnvironment'
 import { useConditionInputType } from 'components/segments/Rule/hooks/useConditionInputType'
