+++ frontend/web/components/segments/Rule/hooks/index.ts
@@ -1,4 +1,3 @@
 export { useRuleOperator } from './useRuleOperator'
 export { useRuleContext } from './useRuleContext'
-export { useConditionCache } from './useConditionCache'
 export { useConditionInputType } from './useConditionInputType'
